requirements.txt
langchain-openai>=0.0.5
python-dotenv>=1.0.0
openai>=1.12.0 